requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
python-dotenv
email-validator